feat(participant): participant 추상클래스 정의하여 player 가 이를 상속
