+++ src/test/java/blackjack/ParticipantTest.java
@@ -17,12 +17,12 @@ import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class PlayerTest {
+public class ParticipantTest {
     private static final Card TWO_DIAMOND = new Card(Number.TWO, Shape.DIAMOND);
     private static final Card JACK_SPADE = new Card(Number.JACK, Shape.SPADE);
     private static final Card THREE_HEART = new Card(Number.THREE, Shape.HEART);
     private static final Card ACE_CLOVER = new Card(Number.ACE, Shape.CLOVER);
-    private Player player;
+    private Participant participant;
 
     private static Stream<Arguments> provideCardsAndExpectedResult() {
         return Stream.of(
@@ -35,42 +35,42 @@ public class PlayerTest {
 
     @BeforeEach
     void setUp() {
-        player = new Player("wannte");
+        participant = new Player("wannte");
     }
 
     @Test
     @DisplayName("플레이어 카드 추가")
     void addCard() {
         Card card = Deck.draw();
-        player.addCard(card);
-        assertThat(player.getCards()).containsExactly(card);
+        participant.addCard(card);
+        assertThat(participant.getCards()).containsExactly(card);
     }
 
     @Test
     void calculateCardsTest() {
-        player.addCard(TWO_DIAMOND);
-        player.addCard(JACK_SPADE);
-        player.addCard(THREE_HEART);
+        participant.addCard(TWO_DIAMOND);
+        participant.addCard(JACK_SPADE);
+        participant.addCard(THREE_HEART);
 
-        assertThat(player.calculateCards()).isEqualTo(15);
+        assertThat(participant.calculateCards()).isEqualTo(15);
     }
 
     @Test
     @DisplayName("ACE 개수 세기")
     void countAce() {
-        player.addCard(ACE_CLOVER);
-        player.addCard(ACE_CLOVER);
-        player.addCard(ACE_CLOVER);
-        assertThat(player.countAce()).isEqualTo(3);
+        participant.addCard(ACE_CLOVER);
+        participant.addCard(ACE_CLOVER);
+        participant.addCard(ACE_CLOVER);
+        assertThat(participant.countAce()).isEqualTo(3);
     }
 
     @ParameterizedTest
     @MethodSource("provideCardsAndExpectedResult")
     void calculateResultTest(List<Card> cards, int expectedResult) {
         for (Card card : cards) {
-            player.addCard(card);
+            participant.addCard(card);
         }
 
-        assertThat(player.calculateResult()).isEqualTo(expectedResult);
+        assertThat(participant.calculateResult()).isEqualTo(expectedResult);
     }
 }
